+++ classes/ThrowableFunctions/ThrowablePHPFunctions.php
@@ -27,7 +27,11 @@ use Webmozart\Assert\InvalidArgumentException;
 function throwOnNull($result)
 {
     if (is_null($result)) {
-        throw new Exception('error (null) ' . debug_backtrace()[1]['function']);
+        if (version_compare(PHP_VERSION, '5.6.0') >= 0) {
+            throw new Exception('error (null) ' . debug_backtrace()[1]['function']);
+        } else {
+            throw new Exception('error (null)');
+        }
     }
     return $result;
 }
@@ -46,7 +50,11 @@ function preg_replace($pattern, $replacement, $subject, $limit = -1, &$count = n
 {
     $result = \preg_replace($pattern, $replacement, $subject, $limit, $count);
     if (is_null($result)) {
-        throw new Exception('error (null) ' . debug_backtrace()[1]['function']);
+        if (version_compare(PHP_VERSION, '5.6.0') >= 0) {
+            throw new Exception('error (null) ' . debug_backtrace()[1]['function']);
+        } else {
+            throw new Exception('error (null)');
+        }
     }
     return $result;
 }